rename mock server to mocker and include utils
