+++ .eslintrc.js
@@ -27,7 +27,7 @@ const config = {
       files: [
         'packages/jest-puppeteer-env/**/*.js',
         'packages/jest-puppeteer-env-example/**/*.js',
-        'packages/jest-puppeteer-mock-server/**/*.js',
+        'packages/mocker/**/*.js',
         'packages/check/**/*.js',
         './.eslintrc.js'
       ],
