+++ packages/jest-puppeteer-env/index.js
@@ -1,3 +1,1 @@
-/* eslint-disable import/group-exports */
 module.exports = require('./src/environment');
-module.exports.mockUtils = require('./src/mockUtils');
