+++ packages/jest-puppeteer-env/package.json
@@ -10,7 +10,8 @@
   "homepage": "https://github.com/chealt/chealt#chealt",
   "scripts": {},
   "dependencies": {
-    "@axe-core/puppeteer": "^4.1.1",
+    "@axe-core/puppeteer": "4.1.1",
+    "@chealt/mocker": "0.0.1",
     "bytes": "3.1.0",
     "chalk": "4.1.0",
     "cwd": "0.10.0",
