+++ packages/jest-puppeteer-env/src/factory.js
@@ -1,4 +1,4 @@
-const { findMocksForUrl } = require('./mockUtils');
+const { findMocksForUrl } = require('@chealt/mocker');
 const { startCollecting, getCoverage } = require('./coverage/index');
 const { writeFileSafe, createDir } = require('./fileUtils');
 const performance = require('./performance/index');
