+++ packages/mocker/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@chealt/jest-puppeteer-mock-server",
+  "name": "@chealt/mocker",
   "version": "0.0.1",
   "license": "MIT",
   "type": "module",
@@ -10,8 +10,8 @@
   },
   "homepage": "https://github.com/chealt/chealt#chealt",
   "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js"
+    "start": "node server.js",
+    "dev": "nodemon server.js"
   },
   "publishConfig": {
     "access": "public",
@@ -22,10 +22,9 @@
     "node": "14"
   },
   "dependencies": {
-    "@chealt/jest-puppeteer-env": "3.0.0",
-    "express": "^4.17.1",
-    "glob": "^7.1.6",
-    "winston": "^3.3.3"
+    "express": "4.17.1",
+    "glob": "7.1.6",
+    "winston": "3.3.3"
   },
   "devDependencies": {
     "nodemon": "^2.0.7"
