+++ packages/mocker/src/utils.js
@@ -28,4 +28,31 @@ const getMocks = async ({ mocksFolderAbsPath, mockExtension }) => {
   return Promise.all(mockFiles.map((mockFilePath) => readFile(mockFilePath).then((buffer) => JSON.parse(buffer))));
 };
 
-export { getFolderAbsPath, getMocks };
+const removePort = (url) => url.replace(/:\d\d\d\d[\d]*/gu, '');
+const removeHost = (url) => url.replace(/https?:\/\/[^/]*/gu, '');
+
+const findMocksForUrl = ({ isPortAgnostic, isHostAgnostic }) => (mocks, url) => {
+  const mockKey = Object.keys(mocks).find((responseUrl) => {
+    if (isPortAgnostic && isHostAgnostic) {
+      return removeHost(removePort(responseUrl)) === removeHost(removePort(url));
+    }
+
+    if (isPortAgnostic) {
+      return removePort(responseUrl) === removePort(url);
+    }
+
+    if (isHostAgnostic) {
+      return removeHost(responseUrl) === removeHost(url);
+    }
+
+    return responseUrl === url;
+  });
+
+  if (mockKey) {
+    return mocks[mockKey];
+  }
+
+  return undefined;
+};
+
+export { getFolderAbsPath, getMocks, findMocksForUrl };
